09._Environments_Variables/python/pyproject.toml
[tool.poetry]
name = "python-dotenv-example"
version = "0.1.0"
description = ""
authors = ["Betül Iskender <betul.iskender@outlook.dk>"]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
